refactor: move types to `lib.rs`
This is so that they can be more easily tested in a REPL.
